+++ src/AGE_library.h
@@ -51,7 +51,7 @@ namespace AGE {
 
 	class ClickTrigger : public Trigger {
 		static constexpr unsigned int DEBOUNCE_DELAY = 10;
-		unsigned int pin;
+		uint8_t pin;
 		ClickTriggerEdge edge;
 		unsigned int sinceLastUp = 0;
